pyproject.toml
[tool.poetry]
name = "api-auth-users"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.2"
pydantic-settings = "^2.6.1"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.36"
sqlalchemy-utils = "^0.41.2"
passlib = "^1.7.4"
python-jose = "^3.3.0"
pytz = "^2024.2"
fastapi = "^0.115.5"
uvicorn = {extras = ["standart"], version = "^0.32.0"}
python-multipart = "^0.0.17"
bcrypt = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
